Rework display scaling around the dpi crate and a central ScaleStore
Adopt the dpi crate for geometry (i32 pixels, bare f64 scale, one
rounding policy via Pixel::from_f64) in place of the homegrown
LogicalSize/PhysicalSize and Scale(f32) newtype.

Add ScaleStore as the sole source of display scale. Platforms are the
exclusive producers: boot probes (Wayland xdg-output, X11 Xft.dpi,
Windows GetDpiForSystem, macOS backingScaleFactor) plus runtime
observers (Wayland fractional-scale callback, X11 RESOURCE_MANAGER
property watch, WM_DPICHANGED in the mpv wndproc hook, macOS
backing-properties notification).

Consumers read the store everywhere. CEF view state carries
logical/physical/scale as one coherent unit (GetScreenInfo no longer
returns a torn pw/w derivation), SurfaceSize and the resize path carry
scale end to end, and the store's changed handler replaces mpv's
display-hidpi-scale observation as the CEF rescale trigger. Menus and
popups drop their per-platform caches.

Unify input coordinate spaces behind InputSpace: platforms declare once
whether pointer coords arrive physical (X11, Windows) or logical
(Wayland, macOS) and dispatch converts to CEF-logical via the active
view's scale — deleting X11's hand-rolled per-event to_logical and
fixing the Windows never-scaled hit-testing bug.

Delete the legacy plumbing: Platform::get_scale/effective_scale
(get_display_scale survives as boot-only probe_display_scale),
per-platform scale caches, the playback scale provider and
display-scale handler.

+++ src/x11/src/scale.rs
@@ -13,7 +13,7 @@ use x11rb::rust_connection::RustConnection;
 
 const BASE_DPI: f64 = 96.0;
 
-pub(crate) fn query_display_scale() -> Option<f32> {
+pub(crate) fn query_display_scale() -> Option<f64> {
     let (conn, screen_num) = RustConnection::connect(None).ok()?;
     if let Some(scale) = query_xft_dpi_scale(&conn) {
         tracing::debug!(target: "x11::scale", "Using Xft.dpi scale: {scale}");
@@ -26,13 +26,13 @@ pub(crate) fn query_display_scale() -> Option<f32> {
     None
 }
 
-fn query_xft_dpi_scale(conn: &impl Connection) -> Option<f32> {
+fn query_xft_dpi_scale(conn: &impl Connection) -> Option<f64> {
     let db = new_from_resource_manager(conn).ok().flatten()?;
     let value: i64 = db.get_value("Xft.dpi", "").ok().flatten()?;
     quantize_dpi(value as f64)
 }
 
-fn query_screen_dpi_scale(conn: &impl Connection, screen_num: usize) -> Option<f32> {
+fn query_screen_dpi_scale(conn: &impl Connection, screen_num: usize) -> Option<f64> {
     let screen = conn.setup().roots.get(screen_num)?;
     let w_mm = screen.width_in_millimeters;
     let h_mm = screen.height_in_millimeters;
@@ -49,15 +49,15 @@ fn query_screen_dpi_scale(conn: &impl Connection, screen_num: usize) -> Option<f
     let sx = quantize_dpi_steps(dpi_x)?;
     let sy = quantize_dpi_steps(dpi_y)?;
     if sx == sy {
-        Some(sx as f32 / 2.0)
+        Some(f64::from(sx) / 2.0)
     } else {
         None
     }
 }
 
-fn quantize_dpi(dpi: f64) -> Option<f32> {
+fn quantize_dpi(dpi: f64) -> Option<f64> {
     let s = quantize_dpi_steps(dpi)?;
-    Some(s as f32 / 2.0)
+    Some(f64::from(s) / 2.0)
 }
 
 fn quantize_dpi_steps(dpi: f64) -> Option<i32> {
